requirements.txt
django==5.2.6
# django-cid==2.3
django-cid==3.0
django-debug-toolbar==4.4.5
django-environ==0.8.1
#Pillow==9.5.0
# psycopg2-binary==2.9.9
# requests==2.32.0
# Install psycopg2-binary==2.9.9 if psycopg==3.2.3 is causing errors.
psycopg==3.2.3
setuptools==65.5.0